# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Campus Event Hub API",
    "version": "1.0.0",
    "description": "Campus Event Hub backend API for managing events, clubs, colleges, and users"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    },
    {
      "url": "https://api.campuseventhub.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Event": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Spring Tech Conference 2025"
          },
          "description": {
            "type": "string",
            "example": "Annual technology conference featuring guest speakers"
          },
          "date": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-25T14:00:00.000Z"
          },
          "venue": {
            "type": "string",
            "example": "Main Auditorium"
          },
          "club": {
            "$ref": "#/components/schemas/ObjectId"
          },
          "attendees": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ObjectId"
            }
          },
          "capacity": {
            "type": "integer",
            "example": 200
          },
          "imageUrl": {
            "type": "string",
            "example": "https://example.com/images/event.jpg"
          }
        }
      },
...
      "ObjectId": {
        "type": "string",
        "description": "MongoDB ObjectId",
        "example": "60f5b2b5c8f7b3c2b8d3e4f5"
      },
...
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "summary": "Get all events",
        "tags": ["Events"],
        "security": [],
        "parameters": [
          {
            "name": "club",
            "in": "query",
            "description": "Filter by club ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search in title/description",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Event"
                      }
                    },
                    "count": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new event (club members only)",
        "tags": ["Events"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "date", "venue", "club", "capacity"],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Spring Tech Conference"
                  },
                  "description": {
                    "type": "string",
                    "example": "Annual technology conference"
                  },
                  "date": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2025-04-25T14:00:00.000Z"
                  },
                  "venue": {
                    "type": "string",
                    "example": "Main Auditorium"
                  },
                  "club": {
                    "type": "string",
                    "description": "Club ObjectId",
                    "example": "60f5b2b5c8f7b3c2b8d3e4f5"
                  },
                  "capacity": {
                    "type": "number",
                    "example": 200
                  },
                  "imageUrl": {
                    "type": "string",
                    "example": "https://example.com/images/event.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created successfully"
          }
        }
      }
    },
...
  }
}